+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/MenuItemBuilder.java
@@ -18,12 +18,7 @@
  */
 package is.codion.swing.common.ui.component.button;
 
-import is.codion.swing.common.ui.control.Control;
-
-import javax.swing.Action;
 import javax.swing.JMenuItem;
-
-import static java.util.Objects.requireNonNull;
 
 /**
  * Builds a JMenuItem.
@@ -36,36 +31,6 @@ public interface MenuItemBuilder<C extends JMenuItem, B extends MenuItemBuilder<
 	 * @return a builder for a JMenuItem
 	 */
 	static <C extends JMenuItem, B extends MenuItemBuilder<C, B>> MenuItemBuilder<C, B> builder() {
-		return new DefaultMenuItemBuilder<>(null);
-	}
-
-	/**
-	 * @param <B> the builder type
-	 * @param <C> the component type
-	 * @param action the button action
-	 * @return a builder for a JButton
-	 */
-	static <C extends JMenuItem, B extends MenuItemBuilder<C, B>> MenuItemBuilder<C, B> builder(Action action) {
-		return new DefaultMenuItemBuilder<>(requireNonNull(action));
-	}
-
-	/**
-	 * @param <B> the builder type
-	 * @param <C> the component type
-	 * @param control the button control
-	 * @return a builder for a JButton
-	 */
-	static <C extends JMenuItem, B extends MenuItemBuilder<C, B>> MenuItemBuilder<C, B> builder(Control control) {
-		return new DefaultMenuItemBuilder<>(requireNonNull(control));
-	}
-
-	/**
-	 * @param <B> the builder type
-	 * @param <C> the component type
-	 * @param controlBuilder the button control builder
-	 * @return a builder for a JButton
-	 */
-	static <C extends JMenuItem, B extends MenuItemBuilder<C, B>> MenuItemBuilder<C, B> builder(Control.Builder<?, ?> controlBuilder) {
-		return new DefaultMenuItemBuilder<>(requireNonNull(controlBuilder).build());
+		return new DefaultMenuItemBuilder<>();
 	}
 }