+++ changelog.md
@@ -92,6 +92,7 @@ Codion Change Log
 - ButtonBuilder, overloaded builder() methods removed.
 - ButtonPanelBuilder, overloaded builder() methods removed.
 - MenuBuilder, overloaded builder() methods removed.
+- MenuItemBuilder, overloaded builder() methods removed.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.EntityTypeBuilder and ConnectionProviderBuilder added.
 ### is.codion.swing.framework.ui